Cast CRUD

+++ src/modules/casts/services/cast.service.js
@@ -1,28 +1,61 @@
 const Cast = require("../models/cast.model");
+const AppError = require("../../../shared/errors/AppError");
 
-const createCast = async (data) => {
-  return await Cast.create(data);
+class CastService {
+
+ createCast = async ({name , image , biography}) => {
+  const castCheck = await Cast.findOne({name})
+  if (castCheck) {
+    throw new AppError("Cast member Already added" , 400)
+  }
+  return await Cast.create({name , image , biography });
 };
 
-const getAllCasts = async () => {
-  return await Cast.find();
+ getAllCasts = async () => {
+  return await Cast.find().sort({ createdAt: -1 });
 };
 
-const getCastById = async (id) => {
+ getCastById = async (id) => {
+  const cast = await Cast.findById(id);
+  if (!cast) {
+    throw new AppError("Cast mamber was not found" , 404)
+  }
   return await Cast.findById(id);
 };
 
-const updateCast = async (id, data) => {
+ updateCast = async (id, data) => {
+    const cast = await Cast.findById(id);
+  if (!cast) {
+    throw new AppError("Cast mamber was not found" , 404)
+  }
+
+    if (data.name) {
+          const existingCast = await Cast.findOne({
+              name: {
+                  $regex: `^${data.name}$`,
+                  $options: "i"
+              },
+          });
+
+          if (existingCast) {
+              throw new AppError("Cast member already exists", 400);
+          }
+      }
   return await Cast.findByIdAndUpdate(id, data, {
     new: true,
   });
 };
 
-const deleteCast = async (id) => {
+ deleteCast = async (id) => {
+  const cast = await Cast.findById(id);
+
+  if (!cast) {
+    throw new AppError("Cast member not found" , 404)
+  }
   return await Cast.findByIdAndDelete(id);
 };
 
-const searchCast = async (name) => {
+ searchCast = async (name) => {
   const searchResults = await Cast.find({
     name: {
       $regex: name,
@@ -32,11 +65,7 @@ const searchCast = async (name) => {
   return searchResults;
 };
 
-module.exports = {
-  createCast,
-  getAllCasts,
-  getCastById,
-  updateCast,
-  deleteCast,
-  searchCast,
-};
+
+}
+
+module.exports = new CastService()
